@@ -1,3 +1,22 @@
+# == Schema Information
+#
+# Table name: sessions
+#
+#  id         :bigint           not null, primary key
+#  token      :string           not null
+#  user_id    :bigint           not null
+#  active     :boolean          default(TRUE), not null
+#  device     :string
+#  browser    :string
+#  location   :string
+#  created_at :datetime         not null
+#  updated_at :datetime         not null
+#
+# Indexes
+#
+#  index_sessions_on_token    (token) UNIQUE
+#  index_sessions_on_user_id  (user_id)
+#
 class Session < ApplicationRecord
   validates :token, presence: true, uniqueness: true
 
@@ -16,6 +35,6 @@ class Session < ApplicationRecord
   end
 
   def deactivate!
-    update_attributes(active: false)
+    update(active: false)
   end
 end
